getdata cron

+++ app/Console/Commands/AutoShutdownCron.php
@@ -5,7 +5,7 @@ namespace App\Console\Commands;
 use Illuminate\Console\Command;
 use Illuminate\Support\Facades\Log;
 use Illuminate\Support\Carbon;
-
+use Illuminate\Support\Facades\DB;
 class AutoShutdownCron extends Command
 {
     /**
@@ -29,8 +29,16 @@ class AutoShutdownCron extends Command
      */
     public function handle()
     {
-        Log::info("Server shutdown at ".now());
-        system('shutdown -h now');
-        return Command::SUCCESS;
+        #Cek cronjob status
+        $cronStatus = DB::table('cronjob')
+                    ->where('cronjob_name','autoshutdown:cron')
+                    ->first();
+
+        if ($cronStatus->cronjob_status == 'open') {
+            info("Server shutdown at ". Carbon::now()->format('Y-m-d H:i:s'));
+            system('shutdown -h now');
+        }else{
+            return Command::SUCCESS;
+        }
     }
 }
